Split the write of Hdf5 files for el-ph tensor in a separate file, for simplicity. Cleaned up the mechanism of selecting the file format.
